Feat: 파일 구조 변경 - 기존에 병합되어있던 클래스들 병합 해제
