+++ marketdata/src/main/java/com/beyond/MKX/domain/indicator/calculator/impl/MomentumCalculator.java
@@ -1,0 +1,46 @@
+package com.beyond.MKX.domain.indicator.calculator.impl;
+
+import com.beyond.MKX.domain.chart.entity.Candle;
+import com.beyond.MKX.domain.indicator.calculator.IndicatorCalculator;
+import com.beyond.MKX.domain.indicator.dto.IndicatorResultDTO;
+import lombok.extern.slf4j.Slf4j;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import java.util.Map;
+
+/** 모멘텀 (Momentum) 계산기 */
+@Slf4j
+class MomentumCalculator implements IndicatorCalculator {
+    private static final int DEFAULT_PERIOD = 10;
+
+    @Override
+    public List<IndicatorResultDTO.IndicatorDataPoint> calculate(List<Candle> candles, Map<String, Object> params) {
+        if (candles == null || candles.isEmpty()) return Collections.emptyList();
+        int period = getIntParam(params, "period", DEFAULT_PERIOD);
+        List<IndicatorResultDTO.IndicatorDataPoint> result = new ArrayList<>();
+
+        for (int i = 0; i < candles.size(); i++) {
+            if (i < period) {
+                result.add(IndicatorResultDTO.IndicatorDataPoint.builder()
+                        .time(candles.get(i).getTime()).values(Map.of("momentum", Double.NaN)).build());
+            } else {
+                double momentum = candles.get(i).getClose() - candles.get(i - period).getClose();
+                result.add(IndicatorResultDTO.IndicatorDataPoint.builder()
+                        .time(candles.get(i).getTime()).values(Map.of("momentum", momentum)).build());
+            }
+        }
+        return result;
+    }
+
+    @Override public Map<String, Object> getDefaultParams() { return Map.of("period", DEFAULT_PERIOD); }
+    @Override public boolean validateParams(Map<String, Object> params) {
+        try { return getIntParam(params, "period", DEFAULT_PERIOD) > 0; } catch (Exception e) { return false; }
+    }
+    private int getIntParam(Map<String, Object> params, String key, int defaultValue) {
+        if (params == null || !params.containsKey(key)) return defaultValue;
+        Object value = params.get(key);
+        return value instanceof Number ? ((Number) value).intValue() : defaultValue;
+    }
+}
